requirements.txt
flask
flask_restx
FlagEmbedding
langchain
langchain-elasticsearch
tiktoken
shortid
jq
elasticsearch==8.12.0
tos
oss2
pyyaml
bson
flask~=3.0.1
Flask-SQLAlchemy~=3.0.5
SQLAlchemy~=1.4.28
psycopg2-binary
pymilvus==2.3.0
qdrant-client==1.7.3
weaviate-client~=3.21.0
redis
flask-migrate~=4.0.5
loguru